deleting image files stored on our oper system when on delete and edit

+++ controllers/admin.js
@@ -1,6 +1,8 @@
 const mongoose = require("mongoose");
-const Product = require("../models/product");
 const { validationResult } = require("express-validator");
+
+const fileHelper = require("../util/file");
+const Product = require("../models/product");
 
 exports.getAddProduct = (req, res, next) => {
   let message = req.flash("error");
@@ -176,6 +178,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then((result) => {
@@ -218,10 +221,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found!"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
-      req.flash("error", "You are not authorized to delete the product!");
+      // req.flash("error", "You are not authorized to delete the product!");
       res.redirect("/admin/products");
     })
     .catch((err) => {
